EES-1838 Move DataBlockCreate/UpdateViewModels into DataBlockViewModels
